+++ packages/slate-tools/src/config.js
@@ -8,7 +8,7 @@ const currentDirectory = __dirname;
 const themeRoot = findRoot(workingDirectory);
 const defaultGulpPath = join(themeRoot, normalize('node_modules/.bin/gulp'));
 // Legacy path for older versions of Node.
-const legacyGulpPath = join(themeRoot, normalize('node_modules/@t-kelly/slate-tools/node_modules/.bin/gulp'));
+const legacyGulpPath = join(themeRoot, normalize('node_modules/t-kelly-slate-tools/node_modules/.bin/gulp'));
 
 const config = {
   gulpFile: join(currentDirectory, 'gulpfile.js'),